% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coherence_panel.R
\encoding{UTF-8}
\name{coherence_panel}
\alias{coherence_panel}
\title{Regional Coherence Index for panel data}
\usage{
coherence_panel(data, geo_dim, kng_dim, kng_nbr, time_dim)
}
\arguments{
\item{data}{It is expected to be a dataframe in "long" format.}

\item{geo_dim}{It is the name of the column of the data.frame that
represents its geographical dimension (e.g., the different regions of
analysis).}

\item{kng_dim}{It is the name of the column of the data.frame that
represents its knowledge dimension (e.g., the different patent classes of
analysis).}

\item{kng_nbr}{It is the name of the column of the data.frame that
represents the numerosity of each knowledge class (e.g., the number of
patents a region has in a given year in a given patent class).}

\item{time_dim}{It is the name of the column of the data.frame that
represents its temporal dimension (e.g., the different years of analysis).}
}
\value{
A data.frame with the Coherence Index of each geographical area
in each time step of analysis.
}
\description{
The function computes the so called Coherence index.
}
\details{
The function computes the so called Coherence index.
It assumes that the "universe" from which you derive the distribution of
reference is composed by all the information provided in the database, and
only this.
You can use it both on a panel data set (if you identify also a column with
a temporal indication of the observations) or on cross-section data (by
leaveing the parameter just said as NULL).
}
\examples{
RCI <- coherence(data = df, time_dim = year, geo_dim = NUTS2,
kng_dim = IPC.3dig, kng_nbr = N.patents)
}
\references{
Teece, Rumelt, Dosi and Winter (1994) ``Understanding Corporate
Coherence: Theory and Evidence'', \emph{Journal of Economic Behavior &
Organization}, 23, 1--30;

Nesta and Saviotti (2005) ``Coherence of the Knowledge Base and the
Firm's Innovative Performance: Evidence from the U.S. Pharmaceutical
Industry'', \emph{Journal of Industrial Economics}, 53, 123-142;

Nesta and Saviotti (2006) 'Firm Knowledge and Market Value in
Biotechnology', \emph{Industrial and Corporate Change}, 15, 625-652;

Bottazzi and Pirino (2010) ``Measuring Industry Relatedness and
Corporate Coherence'', \emph{SSRN Electronic Journal}, 11, 1--24;

Quatraro (2010) 'Knowledge Coherence, Variety and Economic Growth:
Manufacturing Evidence from Italian Regions', \emph{Research Policy}, 39,
1289-1302;

Rocchetta and Mina (2017), ``Technological Coherence and the Adaptive
Resilience of Regional Economies'', \emph{Papers in Evolutionary Economic
Geography}, Utrecht University.
}
