% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{entropy}
\alias{entropy}
\title{Information Entropy}
\usage{
entropy(data)
}
\arguments{
\item{data}{Frequency distribution of a vector of elements.}
}
\value{
The information entropy level of the vector.
}
\description{
The function will return the frequency distribution of a given vector.
It is useful for entropy functions.
}
\details{
The function will return the frequency distribution of a given vector.
It is useful for entropy functions.
}
\examples{
frequences <- get_freqs(c(1,3,5,2,4,9,1,6,10))
etpy <- entropy(frequences)
}
