% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_decomposition_panel.R
\encoding{UTF-8}
\name{entropy_decomposition_panel}
\alias{entropy_decomposition_panel}
\title{Entropy Decomposition Theorem for panel data}
\usage{
entropy_decomposition_panel(data, kng_nbr, kng_dim_upper, geo_dim, time_dim)
}
\arguments{
\item{data}{It is expected to be a dataframe in "long" format.}

\item{kng_nbr}{It is expected to be the name of the (numeric) column of the
DB in which there is either the absolute number of patents (firms) of a
given class (industry), or the corresponding relative frequences.}

\item{kng_dim_upper}{It is expected to be an integer that denotes the number
of digits considered as the level of aggregation to compute the entropy
decomposition (that is, these are the groups to which the word 'within
groups' and 'between groups' refer)}

\item{geo_dim}{It is expected to be a (character) column of the DB that
identifies the first level of aggregation of the data. E.g., can be a
geographical area (county, region, nation) or can refer to the temporal
dimension (year). The name of the column in the output DB can be chosen
through the geo_dim_label option, otherwise a generic label will be use.}

\item{time_dim}{It is expected to be a (character) column of the DB that
identifies the second level of aggregation of the data. It is take for
granted that you cannot use it without geo_dim.}
}
\value{
the value of the entropy index if you provide only an array of
values. Otherwise, a data.table with the overall entropy index, and its
decomposition in between- and within-group components.
}
\description{
The function will return the information entropy of a given vector of
frequences and the same information decomposed in a between- and a
within-groups component. The last two are the ones also respectively called
Unrelated and Related Variety by the Regional Economics of Innovation
literature.
}
\details{
The function will return the so called Entropy Decomposition Theorem
so that, besides the overall entropy level, also the value of the between-
and within-groups information entropy is provided. Internally the function
computes and temorarily stores also other values. However they are not
provided in the oputput dataframe.
}
\examples{
Variety(data = data, kng_nbr = N, geo_dim = region,
time_dim = year, kng_dim_upper = IPC.1,
geo_dim_label = "region", time_dim_label = "year")
Variety(data = data$N)
Variety(data = data, kng_nbr = freq, kng_dim_upper = IPC.3)
}
\references{
Theil (1972) \emph{Statistical Decomposition Analysis}, North-Holland;

Zadjenweber (1972) ``Une Application de la Théorie de l'Information
à l'Économie: La Mesure de la Concentration'', \emph{Revue d'Economie
Politique}, 82, 486--510;

Attaran (1985) ``Industrial Diversity and Economic Performance in U.S.
Areas'', \emph{The Annals of Regional Science}, 20, 44--54;

Frenken (2007) ``Entropy Statistics and Information Theory'', in
Hanusch and Pyka (Eds.) \emph{Elgar Companion to Neo-Schumpeterian
Economics}, Edward Elgar;

Frenken, van Oort and Verburg (2007) ``Related Variety, Unrelated
Variety and Regional Economic Growth'', \emph{Regional Studies}, 41,
685--697;

Boschma and Iammarino (2009) ``Related Variety, Trade Linkages, and Regional
Growth in Italy'', \emph{Economic Geography}, 85, 289--311;

Quatraro (2010) ``Knowledge Coherence, Variety and Economic Growth:
Manufacturing Evidence from Italian Regions'', \emph{Research Policy}, 39,
1289--1302;

#' Content and Frenken (2016) ``Related Variety and Economic Development: A
Literature Review'', \emph{European Planning Studies}, 24, 2097--2112;

Rocchetta and Mina (2017), ``Technological Coherence and the Adaptive
Resilience of Regional Economies'', \emph{Papers in Evolutionary Economic
Geography}, Utrecht University.
}
