% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness.R
\name{fitness}
\alias{fitness}
\title{Regional Knowledge Fitness Index}
\usage{
fitness(data, geo_dim, kng_dim, kng_nbr, time_dim = NULL)
}
\arguments{
\item{data}{It is expected to be a dataframe in "long" format.}

\item{geo_dim}{It is the name of the column of the data.frame that
represents its geographical dimension (e.g., the different regions of
analysis).}

\item{kng_dim}{It is the name of the column of the data.frame that
represents its knowledge dimension (e.g., the different patent classes of
analysis).}

\item{kng_nbr}{It is the name of the column of the data.frame that
represents the numerosity of each knowledge class (e.g., the number of
patents a region has in a given year in a given patent class).}

\item{time_dim}{It is the name of the column of the data.frame that
represents its temporal dimension (e.g., the different years of analysis).
It is an optional parameter}
}
\value{
A data.frame with the Fitness Index of each geographical area
in each time step of analysis.
}
\description{

}
\details{
See:
\itemize{
\item{Tacchella, Cristelli, Caldarelli, Gabrielli and Pietronero (2012)
"A New Metrics for Countries' Fitness and Products' Complexity",
\emph{Scientific Reports}, 2, 1--7;}
\item{Balland and Rigby (2017) "The Geography of Complex Knowledge",
\emph{Economic Geography}, 93, 1--23.}
}
}
\examples{
RKFI <- fitness(data = df, time_dim = year, geo_dim = NUTS2,
kng_dim = IPC.3dig, kng_nbr = N.patents)
}
