% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_freqs.R
\name{get_freqs}
\alias{get_freqs}
\title{Get the frequency distribution of a given vector}
\usage{
get_freqs(data)
}
\arguments{
\item{data}{Vector of numbers. If it sums to 1, it assumes they are already
frequencies.}
}
\value{
A vector of numbers that sums to 1.
}
\description{
The function will return the frequency distribution of a given vector.
It is useful for entropy functions.
}
\details{

}
\examples{
frequences <- get_freqs(values)
frequences <- get_freqs(table(values))
}
