% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard_index.R
\encoding{UTF-8}
\name{jaccard_index}
\alias{jaccard_index}
\title{Jaccard similarity coefficient}
\usage{
jaccard_index(data, geo_dim, kng_dim, kng_nbr, time_dim, binary_mode = "RTA")
}
\description{
The function computes the Jaccard similarity coefficient.
}
\details{
The function computes the Jaccard similarity coefficient.
}
\references{

}
