% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_maximum_spanning_tree.R
\encoding{UTF-8}
\name{plot.reks_maximum_spanning_tree}
\alias{plot.reks_maximum_spanning_tree}
\title{Plot the Maximum Spanning Tree of a proximity matrix}
\usage{
\method{plot}{reks_maximum_spanning_tree}(mst, ...)
}
\arguments{
\item{phi_df}{proximity matrix of type \bold{reks_proximity_hkbh} or
\bold{reks_proximity_zctp}}

\item{time_laps}{Year or vector of years to get a subset of the phi_df
(optional)}

\item{min_thrashold}{All the edges with a weight greater that this thrashold
will be added to the Maximum Spanning Tree}
}
\value{
nothing
}
\description{
It computes the Maximum Spanning Tree for proximity matrices
of class \bold{reks_proximity_hkbh} or \bold{reks_proximity_zctp}, and plot
it.
}
\details{
You can use the functions \code{proximity_hkbh()} or
\code{proximity_zctp()} (and the related ones for panel data) to get a
\code{phi_df} object of the proper class.
}
