% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_biadj_matrix.R
\encoding{UTF-8}
\name{plot.rks_biadj_matrix}
\alias{plot.rks_biadj_matrix}
\title{Plot a biadjacency matrix}
\usage{
\method{plot}{rks_biadj_matrix}(x, order = "DU", ...)
}
\arguments{
\item{x}{It is expected to be an object of class \code{rks_biadj_matrix}.}

\item{order}{It is the way in which rows and columns of the matrix are
re-arranged. By now the only possible option is "DU" that arranges the rows
and columns in order inverse to their degree (i.e., to the
\emph{diversification} of the geographical area and the \emph{ubiquity} of
the knowledge dimension considered.)}

\item{...}{Other parameters taken by plot.}
}
\value{
Nothing
}
\description{
The function plots a biadjacency matrix.
}
\details{
Use \code{\link{plot_biadj_matrix}} instead of this function to have a better
and easier user experience.
}
