% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_biadj_matrix.R
\encoding{UTF-8}
\name{plot_biadj_matrix}
\alias{plot_biadj_matrix}
\title{Plot a biadjacency matrix}
\usage{
plot_biadj_matrix(data, geo_dim, kng_dim, kng_nbr, binary_mode, order = "DU",
  ...)
}
\arguments{
\item{data}{It must be a ``long'' dataframe.}

\item{geo_dim}{It must be a symbolic variable that denotes the name of the
column of the dataframe that represents the geographical units of analysis.}

\item{kng_dim}{It must be a symbolic variable that denotes the name of the
column of the dataframe that represents the technological knowledge units of
analysis.}

\item{kng_nbr}{It must be a symbolic variable that denotes the name of the
column of the dataframe that represents the number of technological units
of a given type that a given geographical units owns.}

\item{binary_mode}{It must be a string that represents the way in which the
data are transformed from count values to binary values. It can be 'none',
'simple', 'RTA', 'RCA', 'higher_quartiles', 'higher_quartiles_kng',
'higher_deciles_kng'.}

\item{order}{It must be a string that represents the way in which the matrix
rows and columns are re-arranged. By now the only possible option is "DU" that arranges the rows
and columns in order inverse to their degree (i.e., to the
\emph{diversification} of the geographical area and the \emph{ubiquity} of
the knowledge dimension considered.)}

\item{...}{Other parameters taken by plot.}
}
\value{
Nothing
}
\description{
The function plots a biadjacency matrix.
}
\details{
The function transforms the ``long'' table provided in a biadjacency matrix
and than plot it.
}
