% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revealed_relatedness.R
\encoding{UTF-8}
\name{proximity}
\alias{proximity}
\title{Proximity}
\usage{
proximity(occt)
}
\arguments{
\item{occt}{Contingency table (i.e., occurrence table or incidence
matrix) on which you want to compute the indices. It can be a 2D array,
in which the first dimension represents the units of analysis (like firms,
regions, or countries), and the second dimension represents the
events or characteristics of interest (like the classes of the patens
produced by the regions, or the sectors in which the workers belongs).
Lastly, the values in each cell represents the occurrences of each unit-event
pair. Moreover, you can use also a 3D array if you like, in which the third
dimension represents the time. The object is expected to be of "table" class.}
}
\value{
A matrix of similarity between technological domains
}
\description{
The function computes the so called Proximity index.
}
\details{
The function computes the so called Proximity index.
}
\references{
Hidalgo, Klinger, Barab{\'a}si, and Hausmann (2007) ``The Product Space
Conditions the Development of Nations'', \emph{Science}.
}
