% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximity_zctp.R
\encoding{UTF-8}
\name{proximity_zctp}
\alias{proximity_zctp}
\title{Proximity a là Zaccaria, Cristelli, Tacchella, Pietronero}
\usage{
proximity_zctp(data, geo_dim, kng_dim, kng_nbr, binary_mode = "RTA",
  names_as_strings = FALSE)
}
\arguments{
\item{data}{It is expected to be a dataframe in "long" format.}

\item{geo_dim}{It is the name of the column of the data.frame that
represents its geographical dimension (e.g., the different regions of
analysis).}

\item{kng_dim}{It is the name of the column of the data.frame that
represents its knowledge dimension (e.g., the different patent classes of
analysis).}

\item{kng_nbr}{It is the name of the column of the data.frame that
represents the numerosity of each knowledge class (e.g., the number of
patents a region has in a given year in a given patent class).}

\item{binary_mode}{It is "RTA" by default. The possible values are: "simple"
(if there is at least a patent in the geo. area in the particular tech. class
it is 1, and 0 otherwise); "RTA" or "RCA" (Balassa method);
"higher_quartiles" (similar to "simple", but it exludes the first quartile
(it seems the best choice, see below); "higher_quartiles_kng" (similar to
'higher_quartiles', but the quartiles are computed for each knowledge class,
assuming that some are more "ubiquitous" than others); "higher_deciles_kng"
(similar to 'higher_quartiles_kng', but it excludes the lower 10% rather than
the lower 25%). Warning: look carefully at the plot_biadj_matrix() output
because I personally observed that with RTA applied to patent data you do not
have the triangular structure observed in the trade data.}

\item{names_as_strings}{If TRUE the function assumes you have written the
name of the columns above said as strings, otherwise they are assumed to
be symbols. Default is FALSE.}
}
\value{
A data.frame with the Coherence Index of each geographical area.
}
\description{
The function computes the so called Proximity, as proposed by Zaccaria,
Cristelli, Tacchella, and Pietronero in 2014 on \emph{PLoS ONE}.
}
\details{

}
\examples{
RCI <- coherence(data = df, geo_dim = NUTS2,
kng_dim = IPC.3dig, kng_nbr = N.patents)
}
\references{
Zaccaria, Cristelli, Tacchella and Pietronero (2014) ``How the Taxonomy of
Products Drives the Economic Development of Countries'', \emph{PLoS ONE},
0113770, 1--17.

Zhou, Ren, Medo and Zhang (2007) ``Bipartite Network Projection and Personal
Recommendation'', \emph{Physical Review E}, 76, 046115.
}
