% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReKS-package.R
\docType{package}
\name{ReKS}
\alias{ReKS}
\alias{ReKS-package}
\title{ReKS: A package for computing some measure about the knowledge space
of a economic region.}
\description{
ReKS: A package for computing some measure about the knowledge space
of a economic region.
}
\section{ReKS main functions}{

\itemize{
\item{\code{\link{shannonEntropy}}}
\item{\code{\link{shannonEvenness}}}
\item{\code{\link{coherence}}}
\item{\code{\link{complexity}}}
\item{\code{\link{fitness}}}
\item{\code{\link{coi}}}
\item{\code{\link{cog}}}
\item{\code{\link{density}}}
}
}

\section{ReKS auxiliary functions}{

\itemize{
\item{\code{\link{data_info}}}
\item{\code{\link{remove_zeros}}}
\item{\code{\link{wide_to_long}}}
\item{\code{\link{rta}}}
\item{\code{\link{relatedness}}}
\item{\code{\link{proximity}}}
\item{\code{\link{local_density}}}
}
}

