% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revealed_relatedness.R
\encoding{UTF-8}
\name{relatedness}
\alias{relatedness}
\title{Relatedness}
\usage{
relatedness(occt, output_statistic = "t", is_binary = NULL,
  fixedmar = "both", seed = Sys.time(), n_sim = 1000,
  sparse = FALSE)
}
\arguments{
\item{occt}{Contingency table (i.e., occurrence table or incidence
matrix) on which you want to compute the indices. It can be a 2D array,
in which the first dimension represents the units of analysis (like firms,
regions, or countries), and the second dimension represents the
events or characteristics of interest (like the classes of the patents
produced by the regions, or the sectors in which the workers belongs).
Lastly, the values in each cell represents the occurrences of each unit-event
pair. Moreover, you can use also a 3D array if you like, in which the third
dimension represents the time. The object is expected to be of "table" class.}

\item{output_statistic}{It can be "t" for t-statistic or "p" for p-value
(look at Bottazzi and Pirino 2010 for an explanation).}

\item{is_binary}{It can be NULL (default), TRUE or FALSE.}

\item{fixedmar}{It is useful to choose which constraint you want to impose
in the null model.}

\item{seed}{It is useful to choose the seed.}

\item{n_sim}{You can choose how many simulations you want to run to compute
the null model.}

\item{sparse}{It can be TRUE or FALSE}
}
\value{
A matrix of similarity between technological domains
}
\description{
The function computes the so called Relatedness index.
}
\details{
The function computes the so called Relatedness index.
}
\references{
Teece, Rumelt, Dosi and Winter (1994) ``Understanding Corporate
Coherence: Theory and Evidence'', \emph{Journal of Economic Behavior &
Organization}, 23, 1--30;

Nesta and Saviotti (2005) ``Coherence of the Knowledge Base and the
Firm's Innovative Performance: Evidence from the {U.S.}~Pharmaceutical
Industry'', \emph{Journal of Industrial Economics}, 53, 123--142;

Nesta and Saviotti (2006) ``Firm Knowledge and Market Value in
Biotechnology'', \emph{Industrial and Corporate Change}, 15, 625--652;

Nesta (2008) ``Knowledge and Productivity in the World's Largest
Manufacturing Corporations'', \emph{Journal of Economic Behavior \&
Organization}, 67, 886--902,

Bottazzi and Pirino (2010) ``Measuring Industry Relatedness and
Corporate Coherence'', \emph{SSRN Electronic Journal}, 11, 1--24;

Quatraro (2010) ``Knowledge Coherence, Variety and Economic Growth:
Manufacturing Evidence from Italian Regions'', \emph{Research Policy}, 39,
1289-1302;
}
