% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_change.R
\name{temporal_change}
\alias{temporal_change}
\title{temporal_change}
\usage{
temporal_change(data, delta = 1)
}
\arguments{
\item{data}{Any onject produced by a ReKS "relatedness measure" function}

\item{delta}{A number or a vector of years (it will return the growth rate of
the measure chosen between a year and the same measure delta years later)}
}
\value{
A data.frame (or a list of data.frames - with a warning in this case)
in which for each delta it has been computed the growth rate between the
measure in a given year and the same measure delta years later.
}
