% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coherence.R
\name{coherence}
\alias{coherence}
\title{Regional Coherence Index}
\usage{
coherence(data, geo_dim, kng_dim, kng_nbr, time_dim = NULL)
}
\arguments{
\item{data}{It is expected to be a data.frame in 'long' format.}

\item{geo_dim}{It is the name of the column of the data.frame that
represents its geographical dimension (e.g., the different regions of
analysis).}

\item{kng_dim}{It is the name of the column of the data.frame that
represents its knowledge dimension (e.g., the different patent classes of
analysis).}

\item{kng_nbr}{It is the name of the column of the data.frame that
represents the numerosity of each knowledge class (e.g., the number of
patents a region has in a given year in a given patent class).}

\item{time_dim}{It is the name of the column of the data.frame that
represents its temporal dimension (e.g., the different years of analysis).
It is an optional parameter}
}
\value{
A data.frame with the Coherence Index of each geographical area
in each time step of analysis.
}
\description{
a
}
\details{
See:
\itemize{
\item{Teece, Rumelt, Dosi and Winter (1994) 'Understanding Corporate
       Coherence: Theory and Evidence', \emph{Journal of Economic
       Behavior \& Organization}, 23, 1-30;}
\item{Nesta and Saviotti (2005) 'Coherence of the Knowledge Base and the
       Firm's Innovative Performance: Evidence from the U.S. Pharmaceutical
       Industry', \emph{Journal of Industrial Economics}, 53, 123-142;}
\item{Nesta and Saviotti (2006) 'Firm Knowledge and Market Value in
       Biotechnology', \emph{Industrial and Corporate Change}, 15, 625-652;}
\item{Quatraro (2010) 'Knowledge Coherence, Variety and Economic Growth:
       Manufacturing Evidence from Italian Regions', \emph{Research Policy},
       39, 1289-1302.}
}
}
\examples{
RCI <- coherence(data = df, time_dim = year, geo_dim = NUTS2,
kng_dim = IPC.3dig, kng_nbr = N.patents)
}
