% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coherence.R
\encoding{UTF-8}
\name{coherence}
\alias{coherence}
\title{Coherence}
\usage{
coherence(occt, relatedness_mtx)
}
\arguments{
\item{occt}{Contingency table (i.e., occurrence table or incidence
matrix) on which you want to compute the indices. It can be a 2D array,
in which the first dimension represents the units of analysis (like firms,
regions, or countries), and the second dimension represents the
events or characteristics of interest (like the classes of the patens
produced by the regions, or the sectors in which the workers belongs).
Lastly, the values in each cell represents the occurrences of each unit-event
pair. Moreover, you can use also a 3D array if you like, in which the third
dimension represents the time. The object is expected to be of "table" class.}

\item{relatedness_mtx}{Matrix of similarity between the technological
classes. The ReKS package provides some function to buid it: see
\code{\link{relatedness}} and \code{\link{proximity}}.}
}
\value{
A data.frame with the Coherence Index of each geographical area.
}
\description{
The function computes the so called Coherence index.
}
\details{
The function computes the so called Coherence index.
It assumes that the "universe" from which you derive the distribution of
reference is composed by all the information provided in the database, and
only this.
You can use it both on a panel data set (if you identify also a column with
a temporal indication of the observations) or on cross-section data (by
leaveing the parameter just said as NULL).
}
\examples{
geo <- paste0("R", 10:50)
tech <- paste0("T", 10:90)
time <- 1:5
dat <- expand.grid(geo, tech, time)
colnames(dat) <- c("geo", "tech", "time")
set.seed(1)
dat$nPat <- sample(1:200, nrow(dat), replace = TRUE)
octab <- xtabs(nPat ~ geo + tech + time, dat)
octab[sample(1:length(octab), length(octab)/2)] <- 0
rel_m <- relatedness(octab, output_statistic = "p")
RCH <- coherence(octab, rel_m)

}
\references{
Teece, Rumelt, Dosi and Winter (1994) ``Understanding Corporate
Coherence: Theory and Evidence'', \emph{Journal of Economic Behavior &
Organization}, 23, 1--30;

Nesta and Saviotti (2005) ``Coherence of the Knowledge Base and the
Firm's Innovative Performance: Evidence from the {U.S.}~Pharmaceutical
Industry'', \emph{Journal of Industrial Economics}, 53, 123--142;

Nesta and Saviotti (2006) ``Firm Knowledge and Market Value in
Biotechnology'', \emph{Industrial and Corporate Change}, 15, 625--652;

Nesta (2008) ``Knowledge and Productivity in the World's Largest
Manufacturing Corporations'', \emph{Journal of Economic Behavior \&
Organization}, 67, 886--902,

Bottazzi and Pirino (2010) ``Measuring Industry Relatedness and
Corporate Coherence'', \emph{SSRN Electronic Journal}, 11, 1--24;

Quatraro (2010) ``Knowledge Coherence, Variety and Economic Growth:
Manufacturing Evidence from Italian Regions'', \emph{Research Policy}, 39,
1289-1302.
}
