% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity.R
\name{complexity}
\alias{complexity}
\title{Regional Knowledge Complexity Index}
\usage{
complexity(data, geo_dim, kng_dim, kng_nbr, time_dim = NULL)
}
\arguments{
\item{data}{It is expected to be a dataframe in "long" format.}

\item{geo_dim}{It is the name of the column of the data.frame that
represents its geographical dimension (e.g., the different regions of
analysis).}

\item{kng_dim}{It is the name of the column of the data.frame that
represents its knowledge dimension (e.g., the different patent classes of
analysis).}

\item{kng_nbr}{It is the name of the column of the data.frame that
represents the numerosity of each knowledge class (e.g., the number of
patents a region has in a given year in a given patent class).}

\item{time_dim}{It is the name of the column of the data.frame that
represents its temporal dimension (e.g., the different years of analysis).
It is an optional parameter}
}
\value{
A data.frame with the Complexity Index of each geographical area
in each time step of analysis.
}
\description{

}
\details{
See:
\itemize{
\item{}
}
}
\examples{
RKCI <- complexity(data = df, time_dim = year, geo_dim = NUTS2,
kng_dim = IPC.3dig, kng_nbr = N.patents)
}
