% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity.R
\encoding{UTF-8}
\name{complexity}
\alias{complexity}
\title{Regional Knowledge Complexity Index}
\usage{
complexity(data, geo_dim, kng_dim, kng_nbr, time_dim = NULL,
  binary_mode = "RTA", scale = TRUE)
}
\arguments{
\item{data}{It is expected to be a dataframe in "long" format.}

\item{geo_dim}{It is the name of the column of the data.frame that
represents its geographical dimension (e.g., the different regions of
analysis).}

\item{kng_dim}{It is the name of the column of the data.frame that
represents its knowledge dimension (e.g., the different patent classes of
analysis).}

\item{kng_nbr}{It is the name of the column of the data.frame that
represents the numerosity of each knowledge class (e.g., the number of
patents a region has in a given year in a given patent class).}

\item{time_dim}{It is the name of the column of the data.frame that
represents its temporal dimension (e.g., the different years of analysis).
It is an optional parameter.}

\item{binary_mode}{It is "RTA" by default. The possible values are: "simple"
(if there is at least a patent in the geo. area in the particular tech. class
it is 1, and 0 otherwise); "RTA" or "RCA" (Balassa method);
"higher_quartiles" (similar to "simple", but it exludes the first quartile
(it seems the best choice, see below); "higher_quartiles_kng" (similar to
'higher_quartiles', but the quartiles are computed for each knowledge class,
assuming that some are more "ubiquitous" than others); "higher_deciles_kng"
(similar to 'higher_quartiles_kng', but it excludes the lower 10% rather than
the lower 25%). Warning: look carefully at the plot_biadj_matrix() output
because I personally observed that with RTA applied to patent data you do not
have the triangular structure observed in the trade data.}

\item{scale}{It is TRUE by default. Otherwise, the Complexity Index is not
standardised (CI - mean[CI] / sd[CI]).}
}
\value{
A data.frame with the Complexity Index of each geographical area
in each time step of analysis.
}
\description{
The function computes the ``Technological Complexity'' of each geographical
area considered, in each given year, in line with the so called \emph{Method
of Reflections} (see the references below).
}
\examples{
RKCI <- complexity(data = df, time_dim = year, geo_dim = NUTS2,
kng_dim = IPC.3dig, kng_nbr = N.patents)
}
\references{
Hidalgo, Klinger, Barabási and Hausmann (2007) ``The Product Space
Conditions the Development of Nations'', \emph{Science}, 317, 482--487;

Hidalgo and Hausmann (2009) ``The Building Blocks of Economic
Complexity'', \emph{PNAS}, 106, 10570--10575;

Antonelli, Crespi, Mongeau Ospina and Scellato (2017) ``Knowledge
Composition, Jacobs Externalities and Innovation Performance in European
Regions'', \emph{Regional Studies}, 51, 1708--1720;

Balland and Rigby (2017) ``The Geography of Complex Knowledge'',
\emph{Economic Geography}, 93, 1--23.
}
