% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_decomposition.R
\encoding{UTF-8}
\name{entropy_decomposition}
\alias{entropy_decomposition}
\title{Entropy Decoposition Theorem}
\usage{
entropy_decomposition(data, groups)
}
\arguments{
\item{data}{Vector of elements.}

\item{groups}{Vector that describes how the elements in the data can be
aggregated in groups.}
}
\value{
A list with four elements. The information entropy level of the vector.
}
\description{
The function will return the information entropy of a given vector of
frequences decomposed in two parts: a between-groups and a within-groups one.
Moreover, it provides you also the probability of each group and the
entropy of each of the groups.
}
\details{
The total (undecomposed) entropy is equal to the sum of the two
decomposed components. The within-groups component is equal to the weighted
average of the entropy of each of the groups. So the between-groups component
is equal to the residual entropy.
}
\examples{
etpy_decomp <- entropy_decomposition(data, grps)
etpy_decomp <- entropy_decomposition(table(data), grps)
}
\references{
Theil (1972) \emph{Statistical Decomposition Analysis}, North-Holland;

Zadjenweber (1972) ``Une Application de la Théorie de l'Information
à l'Économie: La Mesure de la Concentration'', \emph{Revue d'Economie
Politique}, 82, 486--510;

Attaran (1985) ``Industrial Diversity and Economic Performance in U.S.
Areas'', \emph{The Annals of Regional Science}, 20, 44--54;

Frenken (2007) ``Entropy Statistics and Information Theory'', in
Hanusch and Pyka (Eds.) \emph{Elgar Companion to Neo-Schumpeterian
Economics}, Edward Elgar;

Frenken, van Oort and Verburg (2007) ``Related Variety, Unrelated
Variety and Regional Economic Growth'', \emph{Regional Studies}, 41,
685--697;

Quatraro (2010) ``Knowledge Coherence, Variety and Economic Growth:
Manufacturing Evidence from Italian Regions'', \emph{Research Policy}, 39,
1289--1302;

Rocchetta and Mina (2017), ``Technological Coherence and the Adaptive
Resilience of Regional Economies'', \emph{Papers in Evolutionary Economic
Geography}, Utrecht University.
}
