% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness.R
\encoding{UTF-8}
\name{fitness}
\alias{fitness}
\title{Regional Fitness and/or Technological Complexity}
\usage{
fitness(occt, rta = TRUE, binary = TRUE, which = "fitness")
}
\arguments{
\item{occt}{Contingency table (i.e., occurrence table or incidence
matrix) on which you want to compute the indices. It can be a 2D array,
in which the first dimension represents the units of analysis (like firms,
regions, or countries), and the second dimension represents the
events or characteristics of interest (like the classes of the patens
produced by the regions, or the sectors in which the workers belongs).
Lastly, the values in each cell represents the occurrences of each unit-event
pair. Moreover, you can use also a 3D array if you like, in which the third
dimension represents the time. The object is expected to be of "table" class.}

\item{rta}{If TRUE (default) it uses the Revealed Technological Advantages
(RTA) of the original data}

\item{binary}{If TRUE (default) it binarises the RTA matrix
(can be used only together with rta=TRUE).}

\item{which}{It can be one of "fitness" (default), "complexity", "both".
The first returns the fitness of each region; the second returns the
complexity of each technological domain; and the third returns both the
indices.}
}
\value{
A data.frame with the Fitness of each region and/or the Complexity of
each technological domain. If a 3D array is provided as input, it returns the
full panel data.frame.
}
\description{
The function computes the fitness index of each region (i.e.
its \emph{competitiveness}), in line with the methodology proposed by
Pietronero and coauthors.
}
\examples{
geo <- paste0("R", 10:50)
tech <- paste0("T", 10:90)
time <- 1:5
dat <- expand.grid(geo, tech, time)
colnames(dat) <- c("geo", "tech", "time")
set.seed(1)
dat$nPat <- sample(1:200, nrow(dat), replace = TRUE)
octab <- xtabs(nPat ~ geo + tech + time, dat)
octab[sample(1:length(octab), length(octab)/2)] <- 0
FX <- fitness(octab)
attr(FX, "convergence")

}
\references{
Tacchella, Cristelli, Caldarelli, Gabrielli and Pietronero (2012)
``A New Metrics for Countries' Fitness and Products' Complexity'',
\emph{Scientific Reports}, 2, 1--7;

Tacchella, Cristelli, Caldarelli, Gabrielli and Pietronero (2013)
``Economic Complexity: Conceptual Grounding of a New Metrics for Global
Competitiveness'', \emph{Journal of Economic Dynamics and Control}, 37,
1683--1691;

Cristelli, Gabrielli, Tacchella, Caldarelli and Pietronero (2013)
``Measuring the Intangibles: A Metrics for the Economic Complexity of
Countries and Products'', \emph{PLoS ONE}, 8, e70726;
}
