% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness.R
\name{fitness}
\alias{fitness}
\title{Regional Knowledge Fitness Index}
\usage{
fitness(data, geo_dim, kng_dim, kng_nbr, time_dim = NULL,
  binary_mode = "RTA")
}
\arguments{
\item{data}{It is expected to be a dataframe in "long" format.}

\item{geo_dim}{It is the name of the column of the data.frame that
represents its geographical dimension (e.g., the different regions of
analysis).}

\item{kng_dim}{It is the name of the column of the data.frame that
represents its knowledge dimension (e.g., the different patent classes of
analysis).}

\item{kng_nbr}{It is the name of the column of the data.frame that
represents the numerosity of each knowledge class (e.g., the number of
patents a region has in a given year in a given patent class).}

\item{time_dim}{It is the name of the column of the data.frame that
represents its temporal dimension (e.g., the different years of analysis).
It is an optional parameter.}

\item{binary_mode}{It is "RTA" by default. The possible values are: "none"
(weighted matrix, i.e. not binary); "simple" (if there is at least a patent
in the geo. area in the particular tech. class it is 1, and 0 otherwise);
"RTA" or "RCA" (Balassa method); "higher_quartiles" (similar to "simple",
but it exludes the first quartile (it seems the best choice, see below);
"higher_quartiles_kng" (similar to 'higher_quartiles', but the quartiles are
computed for each knowledge class, assuming that some are more "ubiquitous"
than others). Warning: look carefully at the plot_biadj_matrix() output
because I personally observed that with RTA applied to patent data you do not
have the triangular structure observed in the trade data.}
}
\value{
A data.frame with the Fitness Index of each geographical area
in each time step of analysis.
}
\description{

}
\details{
See:
\itemize{
\item{Tacchella, Cristelli, Caldarelli, Gabrielli and Pietronero (2012)
"A New Metrics for Countries' Fitness and Products' Complexity",
\emph{Scientific Reports}, 2, 1--7;}
\item{Tacchella, Cristelli, Caldarelli, Gabrielli and Pietronero (2013)
"Economic Complexity: Conceptual Grounding of a New Metrics for Global
Competitiveness", \emph{Journal of Economic Dynamics and Control}, 37,
1683--1691;}
\item{Cristelli, Gabrielli, Tacchella, Caldarelli and Pietronero (2013)
"Measuring the Intangibles: A Metrics for the Economic Complexity of
Countries and Products", \emph{PLOS ONE}, 8, e70726;}
\item{Balland and Rigby (2017) "The Geography of Complex Knowledge",
\emph{Economic Geography}, 93, 1--23.}
}
}
\examples{
RKFI <- fitness(data = df, time_dim = year, geo_dim = NUTS2,
kng_dim = IPC.3dig, kng_nbr = N.patents)
}
