% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReKS.R
\docType{package}
\name{ReKS}
\alias{ReKS}
\alias{ReKS-package}
\title{ReKS: A package for computating some measure about the knowledge space
of a economic region.}
\description{
ReKS: A package for computating some measure about the knowledge space
of a economic region.
}
\section{ReKS functions}{

\itemize{
\item{\code{\link{entropy}}}
\item{\code{\link{entropy_decomposition}}}
\item{\code{\link{entropy_decomposition_panel}}}
\item{\code{\link{coherence}}}
\item{\code{\link{coherence_panel}}}
\item{\code{\link{complexity}}}
\item{\code{\link{fitness}}}
}
}

